requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
pymongo==4.6.0
prometheus-client==0.18.0
jinja2==3.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.0.0
pydantic[email]==2.4.2